rename package model -> costattributionmodel.
