+++ pkg/costattribution/active_tracker_test.go
@@ -12,13 +12,13 @@ import (
 	"github.com/prometheus/prometheus/model/labels"
 	"github.com/stretchr/testify/assert"
 
-	"github.com/grafana/mimir/pkg/costattribution/model"
+	"github.com/grafana/mimir/pkg/costattribution/costattributionmodel"
 )
 
 func TestActiveTracker_hasSameLabels(t *testing.T) {
 	manager, _, _ := newTestManager()
 	ast := manager.ActiveSeriesTracker("user1")
-	assert.True(t, ast.hasSameLabels([]model.Label{{Input: "team", Output: ""}}), "Expected cost attribution labels mismatch")
+	assert.True(t, ast.hasSameLabels([]costattributionmodel.Label{{Input: "team", Output: ""}}), "Expected cost attribution labels mismatch")
 }
 
 func TestActiveTracker_IncrementDecrement(t *testing.T) {
